wip: export aws/Application.on_create(..) function to make it usable by DO implementation
