+++ src/cloud_provider/aws/application.rs
@@ -1,22 +1,19 @@
 use tera::Context as TeraContext;
 
 use crate::build_platform::Image;
-use crate::cloud_provider::common::kubernetes::{
-    do_stateless_service_cleanup, get_stateless_resource_information,
+use crate::cloud_provider::environment::Environment;
+use crate::cloud_provider::kubernetes::{
+    do_stateless_service_cleanup, get_stateless_resource_information, Kubernetes,
 };
-use crate::cloud_provider::common::models::{
+use crate::cloud_provider::models::{
     EnvironmentVariable, EnvironmentVariableDataTemplate, Storage, StorageDataTemplate,
 };
-use crate::cloud_provider::environment::Environment;
-use crate::cloud_provider::kubernetes::Kubernetes;
 use crate::cloud_provider::service::{
-    Action, Application as CApplication, Create, Delete, Pause, Service, ServiceType,
-    StatelessService,
+    deploy_application, Action, Application as CApplication, Create, Delete, Helm, Pause, Service,
+    ServiceType, StatelessService,
 };
 use crate::cloud_provider::DeploymentTarget;
-use crate::cmd::helm::Timeout;
-use crate::cmd::structs::LabelsContent;
-use crate::error::{cast_simple_error_to_engine_error, EngineError, EngineErrorCause};
+use crate::error::{cast_simple_error_to_engine_error, EngineError};
 use crate::models::Context;
 
 #[derive(Clone, Eq, PartialEq, Hash)]
@@ -67,10 +64,6 @@ impl Application {
             storage,
             environment_variables,
         }
-    }
-
-    fn helm_release_name(&self) -> String {
-        crate::string::cut(format!("application-{}-{}", self.name(), self.id()), 50)
     }
 
     fn context(&self, kubernetes: &dyn Kubernetes, environment: &Environment) -> TeraContext {
@@ -163,6 +156,16 @@ impl crate::cloud_provider::service::Application for Application {
     fn set_image(&mut self, image: Image) {
         self.image = image;
     }
+
+    fn start_timeout_in_seconds(&self) -> u32 {
+        self.start_timeout_in_seconds
+    }
+}
+
+impl Helm for Application {
+    fn helm_release_name(&self) -> String {
+        crate::string::cut(format!("application-{}-{}", self.name(), self.id()), 50)
+    }
 }
 
 impl StatelessService for Application {}
@@ -218,97 +221,9 @@ impl Create for Application {
         };
 
         let context = self.context(kubernetes, environment);
-        let workspace_dir = self.workspace_directory();
+        let charts_dir = format!("{}/aws/charts/q-application", self.context.lib_root_dir());
 
-        let from_dir = format!("{}/aws/charts/q-application", self.context.lib_root_dir());
-
-        let _ = cast_simple_error_to_engine_error(
-            self.engine_error_scope(),
-            self.context.execution_id(),
-            crate::template::generate_and_copy_all_files_into_dir(
-                from_dir.as_str(),
-                workspace_dir.as_str(),
-                &context,
-            ),
-        )?;
-
-        // render
-        // TODO check the rendered files?
-        let helm_release_name = self.helm_release_name();
-
-        let kubernetes_config_file_path = kubernetes.config_file_path()?;
-
-        // define labels to add to namespace
-        let namespace_labels = match self.context.resource_expiration_in_seconds() {
-            Some(_) => Some(vec![
-                (LabelsContent {
-                    name: "ttl".to_string(),
-                    value: format! {"{}", self.context.resource_expiration_in_seconds().unwrap()},
-                }),
-            ]),
-            None => None,
-        };
-
-        // create a namespace with labels if do not exists
-        let _ = cast_simple_error_to_engine_error(
-            self.engine_error_scope(),
-            self.context.execution_id(),
-            crate::cmd::kubectl::kubectl_exec_create_namespace(
-                kubernetes_config_file_path.as_str(),
-                environment.namespace(),
-                namespace_labels,
-                kubernetes
-                    .cloud_provider()
-                    .credentials_environment_variables(),
-            ),
-        )?;
-
-        // do exec helm upgrade and return the last deployment status
-        let helm_history_row = cast_simple_error_to_engine_error(
-            self.engine_error_scope(),
-            self.context.execution_id(),
-            crate::cmd::helm::helm_exec_with_upgrade_history(
-                kubernetes_config_file_path.as_str(),
-                environment.namespace(),
-                helm_release_name.as_str(),
-                workspace_dir.as_str(),
-                Timeout::Value(self.start_timeout_in_seconds),
-                kubernetes
-                    .cloud_provider()
-                    .credentials_environment_variables(),
-            ),
-        )?;
-
-        // check deployment status
-        if helm_history_row.is_none() || !helm_history_row.unwrap().is_successfully_deployed() {
-            return Err(self.engine_error(
-                EngineErrorCause::User(
-                    "Your application didn't start for some reason. \
-                Are you sure your application is correctly running? You can give a try by running \
-                locally `qovery run`. You can also check the application log from the web \
-                interface or the CLI with `qovery log`",
-                ),
-                format!("Application {} has failed to start ⤬", self.name_with_id()),
-            ));
-        }
-
-        // check app status
-        let selector = format!("app={}", self.name());
-
-        let _ = cast_simple_error_to_engine_error(
-            self.engine_error_scope(),
-            self.context.execution_id(),
-            crate::cmd::kubectl::kubectl_exec_is_pod_ready_with_retry(
-                kubernetes_config_file_path.as_str(),
-                environment.namespace(),
-                selector.as_str(),
-                kubernetes
-                    .cloud_provider()
-                    .credentials_environment_variables(),
-            ),
-        )?;
-
-        Ok(())
+        deploy_application(kubernetes, environment, self, charts_dir.as_str(), &context)
     }
 
     fn on_create_check(&self) -> Result<(), EngineError> {
