--- src/cloud_provider/common/mod.rs
@@ -1,2 +1,0 @@
-pub mod kubernetes;
-pub mod models;
