+++ src/cloud_provider/digitalocean/router.rs
@@ -3,10 +3,9 @@ use retry::OperationResult;
 use serde::{Deserialize, Serialize};
 use tera::Context as TeraContext;
 
-use crate::cloud_provider::common::kubernetes::do_stateless_service_cleanup;
 use crate::cloud_provider::digitalocean::DO;
 use crate::cloud_provider::environment::Environment;
-use crate::cloud_provider::kubernetes::Kubernetes;
+use crate::cloud_provider::kubernetes::{do_stateless_service_cleanup, Kubernetes};
 use crate::cloud_provider::service::{
     Action, Create, Delete, Pause, Router as RRouter, Service, ServiceType, StatelessService,
 };
